+++ 제이_서제경/5주차/Umc-Mission2/.env
@@ -1,20 +1,1 @@
-# Environment variables declared in this file are automatically made available to Prisma.
-# See the documentation for more detail: https://pris.ly/d/prisma-schema#accessing-environment-variables-from-the-schema
-
-# Prisma supports the native connection string format for PostgreSQL, MySQL, SQLite, SQL Server, MongoDB and CockroachDB.
-# See the documentation for all the connection string options: https://pris.ly/d/connection-strings
-
-DATABASE_URL="file:./dev.db"
-
-JWT_SECRET=m7uah9NZsAXlZ/uKiMyzvPL6XJmnxFyR2jr/M4dOAFY=
-JWT_EXPIRES_IN=3600s
-#JWT_EXPIRES_IN=3600s
-
-
-REFRESH_JWT_SECRET=JoDpot4aMrtCrtj/qGRmwCQu1m76SlSLnCSCnFJ8eMI=
-REFRESH_JWT_EXPIRES_IN=7d
-
-# Google Auth
-GOOGLE_CLIENT_ID=aaa
-GOOGLE_CLIENT_SECRET=aaa
-GOOGLE_CALLBACK_URL=aaa
+VITE_SERVER_API_URL=http://localhost:8000
