+++ 제이_서제경/5주차/Umc-Mission2/src/pages/LoginPage.tsx
@@ -3,17 +3,10 @@ import { UserSigninInformation, validateSignin } from "../utils/vaildate";
 import useForm from "../hooks/useForm";
 import { useNavigate } from "react-router-dom";
 import { useAuth } from "../context/AuthContext";
-import { useEffect } from "react";
 
 const LoginPage = () => {
   const { login, accessToken } = useAuth();
   const navigate = useNavigate();
-
-  useEffect(() => {
-    if (accessToken) {
-      navigate("/");
-    }
-  }, [navigate, accessToken]);
 
   const { values, errors, touched, getInputProps } =
     useForm<UserSigninInformation>({
